requirements.txt
streamlit
joblib
numpy
pandas
liac-arff
scikit-learn
xgboost
